Make better use of history root view
Most of the time when accessing the history, the current
history is the most important one to categorize tracks, etc.
Currently the user has to open the menu, scroll to the last entry
to access the current playlist. Instead use the root view
as a shortcut to view the current playlist.